Subclass CollectDWorker for async send
CollectDBufferedAsyncWorker subclasses CollectDWorker and overrides
its send method so that all calls to send() put the data in a deque
that is flashed into the actual worker pipe in a separate thread.

This means that when the CollectDServerMP which listens to the socket
wants to send data to a worker, the send call will not block if the
pipe is full. Also, because of the deque buffering, the system can
handle traffic spikes quickly enough without having network buffers
filled and start dropping incoming packets.