Send complex names from collectd to influxdb
Bucky was written with Graphite in mind and Graphite uses a dot
separated metric name to organize time series. InfluxDB is more flexible
in that it allow a time series to be arbitrarily tagged.

In this commit, the bucky server's to bucky client's API is changed in a
backwards compatible way so that a server, instead of sending a string
name, can optionally send a list of key/value tuples instead. Clients
that only support single string names (such as carbon), simply
concatenate the values of the pairs to produce the series name. Clients
that don't require a single string name (such as influxdb) can take
advantage of the key/values structured information about the time
series. The new API is currently only used by CollectDServer and
InfluxDBClient.

Note that this commit is backwards incompatible with existing custom
CollectDConverter's.